fix: optimiz index.js
