+++ src/index.html
@@ -8,7 +8,6 @@
   <title>JavaScript Site-Constructor</title>
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-+0n0xVW2eSR5OomGNYDnhzAbDsOXxcvSN1TPprVMTNDbiYZCxYbOOl7+AMvyTG2x" crossorigin="anonymous">
-  <link rel="stylesheet" href="./main.css">
 </head>
 <body>
 <div id="app" class="container-fluid">
