+++ src/model.js
@@ -1,0 +1,6 @@
+export const model = [
+  {type: 'title', value: 'This title is h1'},
+  {type: 'text', value: 'Some text in our block'},
+  {type: 'columns', value: ['111', '222', '333']},
+  {type: 'image', value: './assets/image.png'}
+];
